move subproblem to single file
